Fix Apply gameplayeffectcontainerspec blueprint function
Added const again to the parameter of the function to fix the blueprint call of the funciton.
Instead, just create a copy of the struct and work with that to set and get values.
Set target type sphere radius to 1.0f instead of 100.0f (better for scalable floats since it will set their multiplier).